blueprint: Refactor to not depend on backendName
Before, we ensured the backend names were consistent in the frontend and
backend configs by passing around the Service we were load balancing,
and using a consistent function to generate the name based on the
Service. This patch simplifies the flow of logic by refactoring the
frontend and backend configs to take the backend name as an argument.
This way, the caller can ensure the names are consistent by simply
passing the same name to both functions.